Update: Major Update Cogwatch
Added cogwatch to the project, this not only removes the clutter of loading all the cogs in main.py, it also auto-reloads every cog whenever a save is made. Also updated agent cog to include a /models command. This will return the current ai and a list of available choices. Finally, admin cog was added. This will be used for general purpose utility for only the owner.